requirements.txt
Django==3.0.4
djangorestframework==3.11.0
requests==2.25.0
drf-yasg==1.17.1
psycopg2-binary==2.8.6
djangorestframework==3.11.0
markdown==3.3.3
djangorestframework-simplejwt==4.4.0